% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PF_sample_1}
\alias{PF_sample_1}
\title{Example Dataset 1}
\format{
A data frame with 100 rows and 4 variables:
\describe{
\item{year}{character cycle identifier}
\item{id}{integer participant ID}
\item{fdcd}{numeric USDA Food Code}
\item{fdwt}{numeric intake weight}
}
}
\usage{
PF_sample_1
}
\description{
The 1st practice dataset for demonstrating the \code{\link{ReFormatDF}} function.
This example only has one cycle, and all the variable formats match the need of classification function.
}
\examples{
data(PF_sample_1)
head(PF_sample_1)
}
\keyword{datasets}
