% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PF_sample_2}
\alias{PF_sample_2}
\title{Example Dataset 2}
\format{
A data frame with 100 rows and 4 variables:
\describe{
\item{year}{character cycle identifier}
\item{id}{integer participant ID}
\item{fdcd}{character USDA Food Code}
\item{fdwt}{numeric intake weight}
}
}
\usage{
PF_sample_2
}
\description{
The 2nd practice dataset for demonstrating the \code{\link{ReFormatDF}} function.
This example only has one cycle, and some variable formats don't match the need of classification function.
}
\examples{
data(PF_sample_2)
head(PF_sample_2)
}
\keyword{datasets}
