% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReFormatDF.R
\name{ReFormatDF}
\alias{ReFormatDF}
\title{Reformat a data frame with recoded cycle values}
\usage{
ReFormatDF(infile, var_cycle, var_Food_code, var_weight, your_cycle, out_cycle)
}
\arguments{
\item{infile}{A data frame containing the variables to be reformatted.}

\item{var_cycle}{Character string, the name of the cycle variable in \code{infile}.}

\item{var_Food_code}{Character string, the name of the food code variable in \code{infile}.}

\item{var_weight}{Character string, the name of the weight variable in \code{infile}.}

\item{your_cycle}{A vector of cycle values to be recoded.}

\item{out_cycle}{A vector of new values that will replace those in \code{your_cycle}.}
}
\value{
A data frame with:
\itemize{
\item \code{cycle} (recoded),
\item \code{Food_code} (numeric),
\item \code{Food_code_weight} (numeric),
and the original cycle/food/weight columns removed.
}
}
\description{
This function takes an input data frame and three variable names
(cycle, food code, weight). It recodes the cycle variable according
to a mapping of \code{your_cycle} → \code{out_cycle}, converts food code and
weight to numeric, removes the old variables, and returns the result.
}
\examples{
data(PF_sample_1)
ReFormatDF(PF_sample_1,
  var_cycle = "year",
  var_Food_code = "fdcd",
  var_weight = "fdwt",
  your_cycle = c(),
  out_cycle = c()
)

data(PF_sample_2)
ReFormatDF(PF_sample_2,
  var_cycle = "year",
  var_Food_code = "fdcd",
  var_weight = "fdwt",
  your_cycle = c("cycle1314"),
  out_cycle = c("1314")
)

data(PF_sample_3)
ReFormatDF(PF_sample_3,
  var_cycle = "year",
  var_Food_code = "fdcd",
  var_weight = "fdwt",
  your_cycle = c("cycle0708", "cycle0910", "cycle1112"),
  out_cycle = c("0708", "0910", "1112")
)

}
